.gitignore: merge back changes from Cockpit
In machines/podman and Cockpit we sorted and re-organised the gitignore
file. In navigator I noticed that starter-kit did not have these changes
yet and so we had to re-apply them so let's fix it in the source.

(cherry picked from commit 471073a2c4aad41233f8b3c4b412ba94efa161f4)